# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DWLite Open API",
        "description": "\u7403\u978b\u4f9b\u5e94\u94fe\u8d44\u6e90\u64ae\u5408\u4e0e\u6e20\u9053\u5206\u9500\u5e73\u53f0 Open API\n\n## \u8ba4\u8bc1\u65b9\u5f0f\n\n\u6240\u6709 API \u8bf7\u6c42\u9700\u8981\u4ee5\u4e0b Header:\n- `X-Api-Key`: API Key ID\n- `X-Timestamp`: Unix \u65f6\u95f4\u6233\uff08\u79d2\uff09\n- `X-Nonce`: \u552f\u4e00 UUID\uff08\u9632\u91cd\u653e\uff09\n- `X-Signature`: HMAC-SHA256 \u7b7e\u540d\n\n## \u7b7e\u540d\u7b97\u6cd5\n\n```\ncanonical = METHOD + \"\\n\" + PATH + \"\\n\" + TIMESTAMP + \"\\n\" + NONCE + \"\\n\" + SHA256(BODY)\nsignature = \"sha256=\" + HMAC-SHA256(canonical, API_SECRET)\n```\n\n## \u54cd\u5e94\u683c\u5f0f\n\n\u6210\u529f\u54cd\u5e94:\n```json\n{\n    \"success\": true,\n    \"data\": { ... },\n    \"requestId\": \"uuid\"\n}\n```\n\n\u9519\u8bef\u54cd\u5e94:\n```json\n{\n    \"success\": false,\n    \"error\": {\n        \"code\": \"ERROR_CODE\",\n        \"message\": \"\u9519\u8bef\u63cf\u8ff0\"\n    },\n    \"requestId\": \"uuid\"\n}\n```\n",
        "contact": {
            "name": "DWLite API Support"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1/open",
            "description": "Open API"
        }
    ],
    "paths": {
...
        "/api/v1/open/merchant/inbound/orders": {
            "get": {
                "tags": [
                    "Merchant - Inbound"
                ],
                "summary": "List inbound orders.",
                "operationId": "get_open_api_merchant_inbound_list",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "draft",
                                "pending",
                                "shipped",
                                "arrived",
                                "receiving",
                                "completed",
                                "partial_completed",
                                "cancelled"
                            ],
                            "nullable": true
                        }
                    },
                    {
                        "name": "createdFrom",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    {
                        "name": "createdTo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    {
                        "name": "orderNo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 50,
                            "nullable": true
                        }
                    },
                    {
                        "name": "warehouseCode",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 50,
                            "nullable": true
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/open/merchant/inbound/orders/{orderNo}": {
            "get": {
                "tags": [
                    "Merchant - Inbound"
                ],
                "summary": "Get inbound order details.",
                "operationId": "get_open_api_merchant_inbound_details",
                "parameters": [
                    {
                        "name": "orderNo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
        "/api/v1/open/warehouse/inbound/orders": {
            "get": {
                "tags": [
                    "Warehouse - Inbound"
                ],
                "summary": "List inbound orders.",
                "operationId": "get_open_api_warehouse_inbound_list",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "in_transit",
                                "arrived",
                                "receiving",
                                "completed",
                                "exception"
                            ],
                            "nullable": true
                        }
                    },
                    {
                        "name": "createdFrom",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    {
                        "name": "createdTo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    {
                        "name": "orderNo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 50,
                            "nullable": true
                        }
                    },
                    {
                        "name": "merchantCode",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 50,
                            "nullable": true
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/open/warehouse/inbound/orders/{orderNo}": {
            "get": {
                "tags": [
                    "Warehouse - Inbound"
                ],
                "summary": "Get inbound order details.",
                "operationId": "get_open_api_warehouse_inbound_details",
                "parameters": [
                    {
                        "name": "orderNo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
        "/api/v1/open/warehouse/inbound/orders/{orderNo}/receive": {
            "post": {
                "tags": [
                    "Warehouse - Inbound"
                ],
                "summary": "Complete receiving of inbound order.",
                "operationId": "post_open_api_warehouse_inbound_receive",
                "parameters": [
                    {
                        "name": "orderNo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReceiveInboundRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/open/warehouse/inbound/orders/{orderNo}/exceptions": {
            "post": {
                "tags": [
                    "Warehouse - Inbound"
                ],
                "summary": "Report inbound exception.",
                "operationId": "post_open_api_warehouse_inbound_report_exception",
                "parameters": [
                    {
                        "name": "orderNo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReportExceptionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ReceiveInboundItemDto": {
                "required": [
                    "sku",
                    "receivedQuantity"
                ],
                "properties": {
                    "sku": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "receivedQuantity": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "shelvingLocation": {
                        "type": "string",
                        "maxLength": 200,
                        "nullable": true
                    },
                    "notes": {
                        "type": "string",
                        "maxLength": 500,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ReceiveInboundRequest": {
                "required": [
                    "items",
                    "receivedAt"
                ],
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ReceiveInboundItemDto"
                        },
                        "minItems": 1
                    },
                    "receivedAt": {
                        "type": "string"
                    },
                    "notes": {
                        "type": "string",
                        "maxLength": 500,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ReportExceptionRequest": {
                "required": [
                    "type",
                    "description",
                    "reportedAt"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "damaged",
                            "missing",
                            "wrong_item",
                            "quantity_mismatch",
                            "quality_issue",
                            "other"
                        ]
                    },
                    "sku": {
                        "type": "string",
                        "maxLength": 100,
                        "nullable": true
                    },
                    "affectedQuantity": {
                        "type": "integer",
                        "minimum": 0,
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 1000
                    },
                    "photoUrls": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "default": []
                    },
                    "reportedAt": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API Key ID",
                "name": "X-Api-Key",
                "in": "header"
            },
            "TimestampAuth": {
                "type": "apiKey",
                "description": "Unix \u65f6\u95f4\u6233\uff08\u79d2\uff09",
                "name": "X-Timestamp",
                "in": "header"
            },
            "NonceAuth": {
                "type": "apiKey",
                "description": "\u552f\u4e00 UUID\uff08\u9632\u91cd\u653e\u653b\u51fb\uff09",
                "name": "X-Nonce",
                "in": "header"
            },
            "SignatureAuth": {
                "type": "apiKey",
                "description": "HMAC-SHA256 \u7b7e\u540d",
                "name": "X-Signature",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "ApiKeyAuth": [],
            "TimestampAuth": [],
            "NonceAuth": [],
            "SignatureAuth": []
        }
    ],
...
}